requirements/base.txt
django==3.2.6
django-environ==0.4.5
django-extensions==3.0.9
django-model-utils==4.0.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
werkzeug==2.0.1
python-stdnum==1.16
pycountry==20.7.3
factory-boy==3.2.0
pytest-django==4.4.0
drf-yasg==1.20.0
elasticsearch==5.5.3
django-haystack[elasticsearch]
drf-haystack
django-cors-headers
